+++ src/Money/IDomainFacade.cs
@@ -122,6 +122,6 @@ namespace Money
         /// <param name="validFrom">A date from which the exchange rate is valid.</param>
         /// <param name="rate">An exchange rate.</param>
         /// <returns>Continuation task.</returns>
-        Task SetExchangeRateAsync(string sourceName, string targetName, DateTime validFrom, decimal rate);
+        Task SetExchangeRateAsync(string sourceName, string targetName, DateTime validFrom, double rate);
     }
 }
